// src/pos/enums/unit.enum.ts
export enum RetailUnit {
  PIECE = 'PIECE',       // ชิ้น/อัน
  SACHET = 'SACHET',     // ซอง
  BOTTLE = 'BOTTLE',     // ขวด
  CAN = 'CAN',           // กระป๋อง
  CUP = 'CUP',           // ถ้วย
  BOX = 'BOX',           // กล่องย่อย
  PACK = 'PACK',         // แพ็ก
  DOZEN = 'DOZEN',       // โหล
  CARTON = 'CARTON',     // ลังกระดาษ/กล่องใหญ่
  CRATE = 'CRATE',       // ลังพลาสติก/ลังขวด
}